DynamicHTML: Fix leaking components when content changes
When ngOnChanges is called, because for instance the content input of
the dynamic-html element has changed, a new rendering is created. It
results on a new DynamicHTMLRef instance replacing the old one, but the
old one was never disposed of, which resulted in the leak of the
previous components, and ngOnDestroy() was never called on them.

This commit fixes that by explicitely calling DynamicHTMLRef.destroy()
before rendering the new content.