DynamicHTML: fix handling of imbricated components
If a component known by DynamicHTMLModule is embedded within another
known component within a <dynamic-html> content, what will happen
(depending on the declaration order of components) is:

  1. a component will be created for the inner element
  2. a component will be created for the outer element
  3. the (post-rendering) innerHTML will be passed to the outer
     component, but any behaviour will be lost

The correct way to handle this case is to use a <dynamic-html> tag again
within the outer component, but since it is the post-rendering innerHTML
that is passed to it, it means the inner component will likely not
recognize the innerHTML that gets passed to it. Also it gets rendered
twice, which is useless.

This commit fixes that behaviour by ensuring a component selector is not
embedded in another known component before creating it. It prevents the
instantiation of extra useless components and fixes using <dynamic-html>
within <dynamic-html>.